Move functions from KDOP, Segment, Tetrahedron to corresponding geometry header files
