+++ src/geometry/ArborX_KDOP.hpp
@@ -202,199 +202,4 @@ struct ArborX::GeometryTraits::coordinate_type<
   using type = Coordinate;
 };
 
-namespace ArborX::Details::Dispatch
-{
-
-template <typename KDOP1, typename KDOP2>
-struct expand<KDOPTag, KDOPTag, KDOP1, KDOP2>
-{
-  KOKKOS_FUNCTION static void apply(KDOP1 &that, KDOP2 const &other)
-  {
-    using Kokkos::max;
-    using Kokkos::min;
-
-    constexpr int n_directions = KDOP1::n_directions;
-    static_assert(KDOP2::n_directions == n_directions);
-    for (int i = 0; i < n_directions; ++i)
-    {
-      that._min_values[i] = min(that._min_values[i], other._min_values[i]);
-      that._max_values[i] = max(that._max_values[i], other._max_values[i]);
-    }
-  }
-};
-
-template <typename KDOP, typename Point>
-struct expand<KDOPTag, PointTag, KDOP, Point>
-{
-  KOKKOS_FUNCTION static void apply(KDOP &kdop, Point const &point)
-  {
-    using Kokkos::max;
-    using Kokkos::min;
-
-    constexpr int DIM = GeometryTraits::dimension_v<Point>;
-    constexpr int n_directions = KDOP::n_directions;
-    for (int i = 0; i < n_directions; ++i)
-    {
-      auto const &dir = kdop.directions()[i];
-      auto proj_i = point[0] * dir[0];
-      for (int d = 1; d < DIM; ++d)
-        proj_i += point[d] * dir[d];
-
-      kdop._min_values[i] = min(kdop._min_values[i], proj_i);
-      kdop._max_values[i] = max(kdop._max_values[i], proj_i);
-    }
-  }
-};
-
-template <typename KDOP, typename Box>
-struct expand<KDOPTag, BoxTag, KDOP, Box>
-{
-  KOKKOS_FUNCTION static void apply(KDOP &kdop, Box const &box)
-  {
-    constexpr int DIM = GeometryTraits::dimension_v<KDOP>;
-    static_assert(DIM == 2 || DIM == 3);
-
-    // NOTE if any of the ranges is invalid, the code below would actually
-    // expand the KDOP which is not what we want.
-    // We may revisit this later and decide passing a valid box becomes a
-    // precondition but this would be a breaking change (going from a wide to a
-    // narrow contract).
-    for (int d = 0; d < DIM; ++d)
-      if (box.minCorner()[d] > box.maxCorner()[d])
-        return;
-
-    using Point = std::decay_t<decltype(box.minCorner())>;
-    if constexpr (DIM == 3)
-    {
-      auto const xmin = box.minCorner()[0];
-      auto const ymin = box.minCorner()[1];
-      auto const zmin = box.minCorner()[2];
-      auto const xmax = box.maxCorner()[0];
-      auto const ymax = box.maxCorner()[1];
-      auto const zmax = box.maxCorner()[2];
-      for (auto const &point : {
-               Point{xmin, ymin, zmin},
-               Point{xmin, ymax, zmin},
-               Point{xmin, ymin, zmax},
-               Point{xmin, ymax, zmax},
-               Point{xmax, ymin, zmin},
-               Point{xmax, ymax, zmin},
-               Point{xmax, ymin, zmax},
-               Point{xmax, ymax, zmax},
-           })
-      {
-        Details::expand(kdop, point);
-      }
-    }
-    else
-    {
-      auto const xmin = box.minCorner()[0];
-      auto const ymin = box.minCorner()[1];
-      auto const xmax = box.maxCorner()[0];
-      auto const ymax = box.maxCorner()[1];
-      for (auto const &point : {
-               Point{xmin, ymin},
-               Point{xmin, ymax},
-               Point{xmax, ymin},
-               Point{xmax, ymax},
-           })
-      {
-        Details::expand(kdop, point);
-      }
-    }
-  }
-};
-
-template <typename Box, typename KDOP>
-struct expand<BoxTag, KDOPTag, Box, KDOP>
-{
-  KOKKOS_FUNCTION static void apply(Box &box, KDOP const &kdop)
-  {
-    constexpr int DIM = GeometryTraits::dimension_v<KDOP>;
-
-    // WARNING implicit requirement on KDOP first DIM directions
-    Box other;
-    for (int d = 0; d < DIM; ++d)
-    {
-      other.minCorner()[d] = kdop._min_values[d];
-      other.maxCorner()[d] = kdop._max_values[d];
-    }
-    Details::expand(box, other);
-  }
-};
-
-template <typename KDOP, typename Box>
-struct intersects<KDOPTag, BoxTag, KDOP, Box>
-{
-  KOKKOS_FUNCTION static constexpr bool apply(KDOP const &kdop, Box const &box)
-  {
-    KDOP other{};
-    Details::expand(other, box);
-    return Details::intersects(kdop, other);
-  }
-};
-
-template <typename Box, typename KDOP>
-struct intersects<BoxTag, KDOPTag, Box, KDOP>
-{
-  KOKKOS_FUNCTION static constexpr bool apply(Box const &box, KDOP const &kdop)
-  {
-    return Details::intersects(kdop, box);
-  }
-};
-
-template <typename Point, typename KDOP>
-struct intersects<PointTag, KDOPTag, Point, KDOP>
-{
-  KOKKOS_FUNCTION static constexpr bool apply(Point const &point,
-                                              KDOP const &kdop)
-  {
-    constexpr int DIM = GeometryTraits::dimension_v<Point>;
-    constexpr int n_directions = KDOP::n_directions;
-    for (int i = 0; i < n_directions; ++i)
-    {
-      auto const &dir = kdop.directions()[i];
-      auto proj_i = point[0] * dir[0];
-      for (int d = 1; d < DIM; ++d)
-        proj_i += point[d] * dir[d];
-
-      if (proj_i < kdop._min_values[i] || proj_i > kdop._max_values[i])
-        return false;
-    }
-    return true;
-  }
-};
-
-template <typename KDOP, typename Point>
-struct intersects<KDOPTag, PointTag, KDOP, Point>
-{
-  KOKKOS_FUNCTION static constexpr bool apply(KDOP const &kdop,
-                                              Point const &point)
-  {
-    return Details::intersects(point, kdop);
-  }
-};
-
-template <typename KDOP1, typename KDOP2>
-struct intersects<KDOPTag, KDOPTag, KDOP1, KDOP2>
-{
-  KOKKOS_FUNCTION static constexpr bool apply(KDOP1 const &kdop,
-                                              KDOP2 const &other)
-  {
-    constexpr int n_directions = KDOP1::n_directions;
-    static_assert(KDOP2::n_directions == n_directions);
-    for (int i = 0; i < kdop.n_directions; ++i)
-    {
-      if (other._max_values[i] < kdop._min_values[i] ||
-          other._min_values[i] > kdop._max_values[i])
-      {
-        return false;
-      }
-    }
-    return true;
-  }
-};
-
-} // namespace ArborX::Details::Dispatch
-
 #endif
