+++ src/geometry/ArborX_Segment.hpp
@@ -65,66 +65,4 @@ struct ArborX::GeometryTraits::coordinate_type<
   using type = Coordinate;
 };
 
-namespace ArborX::Details::Dispatch
-{
-
-using GeometryTraits::BoxTag;
-using GeometryTraits::PointTag;
-using GeometryTraits::SegmentTag;
-
-template <typename Box, typename Segment>
-struct expand<BoxTag, SegmentTag, Box, Segment>
-{
-  KOKKOS_FUNCTION static void apply(Box &box, Segment const &segment)
-  {
-    Details::expand(box, segment._start);
-    Details::expand(box, segment._end);
-  }
-};
-
-template <typename Point, typename Segment>
-struct distance<PointTag, SegmentTag, Point, Segment>
-{
-  KOKKOS_FUNCTION static auto apply(Point const &point, Segment const &segment)
-  {
-    constexpr int DIM = GeometryTraits::dimension_v<Point>;
-    using Coordinate = GeometryTraits::coordinate_type_t<Point>;
-
-    if (Details::equals(segment._start, segment._end))
-      return Details::distance(point, segment._start);
-
-    auto const dir = segment._end - segment._start;
-
-    // The line of the segment [a,b] is parametrized as a + t * (b - a).
-    // Find the projection of the point to that line, and clamp it.
-    auto t =
-        Kokkos::clamp(dir.dot(point - segment._start) / dir.dot(dir),
-                      static_cast<Coordinate>(0), static_cast<Coordinate>(1));
-
-    Point projection;
-    for (int d = 0; d < DIM; ++d)
-      projection[d] = segment._start[d] + t * dir[d];
-
-    return Details::distance(point, projection);
-  }
-};
-
-template <typename Segment>
-struct centroid<SegmentTag, Segment>
-{
-  KOKKOS_FUNCTION static auto apply(Segment const &segment)
-  {
-    constexpr int DIM = GeometryTraits::dimension_v<Segment>;
-    using Coordinate = GeometryTraits::coordinate_type_t<Segment>;
-
-    // WARNING implicit requirement on KDOP first DIM directions
-    Point<DIM, Coordinate> point;
-    for (int d = 0; d < DIM; ++d)
-      point[d] = (segment._start[d] + segment._end[d]) / 2;
-    return point;
-  }
-};
-
-} // namespace ArborX::Details::Dispatch
-
 #endif
