+++ src/geometry/ArborX_Tetrahedron.hpp
@@ -55,52 +55,4 @@ struct ArborX::GeometryTraits::coordinate_type<
   using type = Coordinate;
 };
 
-namespace ArborX::Details::Dispatch
-{
-using GeometryTraits::BoxTag;
-using GeometryTraits::PointTag;
-using GeometryTraits::TetrahedronTag;
-
-// expand a box to include a tetrahedron
-template <typename Box, typename Tetrahedron>
-struct expand<BoxTag, TetrahedronTag, Box, Tetrahedron>
-{
-  KOKKOS_FUNCTION static void apply(Box &box, Tetrahedron const &tet)
-  {
-    Details::expand(box, tet.a);
-    Details::expand(box, tet.b);
-    Details::expand(box, tet.c);
-    Details::expand(box, tet.d);
-  }
-};
-
-template <typename Point, typename Tetrahedron>
-struct intersects<PointTag, TetrahedronTag, Point, Tetrahedron>
-{
-  KOKKOS_FUNCTION static constexpr bool apply(Point const &point,
-                                              Tetrahedron const &tet)
-  {
-    static_assert(GeometryTraits::dimension_v<Point> == 3);
-
-    constexpr int N = 4;
-    Kokkos::Array<decltype(tet.a), N> v = {tet.a, tet.b, tet.c, tet.d};
-
-    // For every plane check that the vertex lies within the same halfspace as
-    // the other tetrahedron vertex. This is a simple but possibly not very
-    // efficient algorithm.
-    for (int j = 0; j < N; ++j)
-    {
-      auto normal = (v[(j + 1) % N] - v[j]).cross(v[(j + 2) % N] - v[j]);
-
-      bool same_half_space =
-          (normal.dot(v[(j + 3) % N] - v[j]) * normal.dot(point - v[j]) >= 0);
-      if (!same_half_space)
-        return false;
-    }
-    return true;
-  }
-};
-
-} // namespace ArborX::Details::Dispatch
-
 #endif
